requirements.txt
# Core Dependencies
PySide6==6.5.2
PyYAML==6.0.1
pandas==2.1.0

# Whisper v3 für Audio-Transkription
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python==0.2.0

# API Support (für Kimi K2)
httpx==0.25.0
python-dotenv==1.0.0

# Utils
plotly==5.17.0
numpy>=1.24.0
more-itertools==10.1.0
